rename: 추천챗봇 (chat -> search)로 이동 및 환경 변수 추가
