+++ src/app/api/ai/index.ts
@@ -1,3 +1,3 @@
 export * from "./chatbot";
-// export * from "./search";
+export * from "./search";
 export * from "./video";
